[webui] Transfer specs from main to status messages controller
